[周转筐] 修改siteCode为int

+++ jy-basic-api/src/main/java/com/jdl/basic/api/domain/akbox/AkboxConfig.java
@@ -17,7 +17,7 @@ public class AkboxConfig implements Serializable {
     private String areaHubCode;
     private String areaHubName;
 
-    private Long siteCode;
+    private Integer siteCode;
     private String siteName;
 
     private Integer smallStock;
