+++ jy-basic-api/src/main/java/com/jdl/basic/api/service/akbox/AkboxConfigJsfService.java
@@ -37,5 +37,5 @@ public interface AkboxConfigJsfService {
      */
     Result<Pager<AkboxConfig>> queryPageByCondition(AkboxConfigQuery query);
 
-    Result<AkboxConfig> queryBySiteCode(Long siteCode);
+    Result<AkboxConfig> queryBySiteCode(Integer siteCode);
 }
